Moved to MemMang
Very good idea, my first intention.
